# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: StorageHub API
  version: 1.0.0
  description: >
    REST API of the StorageHub self-service storage management system.
    All endpoints live under the `/api/v1` prefix. Everything except the
    explicitly public endpoints requires a JWT Bearer token (AD-5).
servers:
  - url: /api/v1
    description: Same-origin (Vite dev proxy / SPA-served deployment)

# Default security: JWT everywhere. Public endpoints (e.g. POST /auth/login)
# override this with `security: []` when they are added.
security:
  - bearerAuth: []
paths:
...
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [Contracts]
      summary: Get contract details by ID
      description: >
        Returns contract metadata, status, policy version, and locked content snapshot terms.
        Accessible by the reservation customer or staff/admin.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Contract ID.
      responses:
        '200':
          description: Contract details with locked terms snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/reservation/{reservationId}:
    get:
      operationId: getContractByReservation
      tags: [Contracts]
      summary: Get latest contract by reservation ID
      description: >
        Returns the latest active contract associated with a reservation.
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reservation ID.
      responses:
        '200':
          description: Latest contract for reservation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{id}/re-draft:
    post:
      operationId: reDraftContract
      tags: [Contracts]
      summary: Staff re-draft contract with new revision chaining
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_ADMIN.
        Atomically marks the current contract as SUPERSEDED (isLatest = 0) and creates
        a new DRAFT contract (isLatest = 1) referencing the superseded contract.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Contract ID to supersede and re-draft.
      responses:
        '200':
          description: New draft contract created and previous contract superseded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{id}/print:
    post:
      operationId: printContract
      tags: [Contracts]
      summary: Mark contract as printed
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_ADMIN.
        Transitions contract status from DRAFT to PRINTED.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Contract ID.
      responses:
        '200':
          description: Contract marked as printed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /contracts/{id}/sign:
    post:
      operationId: signContract
      tags: [Contracts]
      summary: Attach signed contract photo and transition status to SIGNED
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_ADMIN.
        Records signed photo URL relative path and transitions contract status to SIGNED.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Contract ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignContractRequest'
      responses:
        '200':
          description: Contract successfully signed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Contract already signed or in invalid status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/reservation/{reservationId}/chain:
    get:
      operationId: getContractChain
      tags: [Contracts]
      summary: Get full contract revision chain for a reservation
      description: >
        Returns all contract versions (including superseded and latest signed/draft) for a reservation.
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reservation ID.
      responses:
        '200':
          description: List of contracts in revision order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContractDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------- attachments (AD-10)
  /attachments:
    post:
      operationId: uploadAttachment
      tags: [Attachments]
      summary: Upload file attachment (signed contract photos, inspection images)
      description: >
        Multipart upload storing files locally and returning relative API path.
        Restricted to authenticated users.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File payload (JPEG, PNG, WebP, PDF up to 10MB).
      responses:
        '201':
          description: File uploaded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentUploadResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /attachments/{filename}:
    get:
      operationId: getAttachment
      tags: [Attachments]
      summary: Stream authenticated file attachment
      description: >
        Authenticated streaming of file attachment. Unauthenticated requests are rejected with 401.
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Attachment filename.
      responses:
        '200':
          description: File stream.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------- tasks (3.2)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT with role claim, 24h TTL, no refresh token (AD-5). Logout is the
        client dropping the token.

  # D6 reusable non-2xx bodies (deferred from 1.1, landed with the first
  # auth paths in 1.3): every usage renders the same Error envelope.
  responses:
    NotFound:
      description: >
        Resource was not found. Error envelope, code NOT_FOUND (404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: >
        Missing, invalid or expired Bearer token - or bad credentials /
        inactive / locked account at login. Error envelope, code
        UNAUTHENTICATED (401). Carries `WWW-Authenticate: Bearer` on guarded
        paths. At login every failure branch shares one generic message.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >
        Authenticated but the token's role is not allowed on this endpoint
        (server-side permission matrix). Error envelope, code FORBIDDEN (403).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: >
        Request body failed validation (missing/invalid fields, duplicate
        email, password confirmation mismatch, terms not accepted). Error
        envelope, code VALIDATION_FAILED (400) with one fieldError per
        offending field.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      description: >
        The one error envelope for every non-2xx response, including security
        (401/403) and business-rule blocks (409). The FE renders toasts from
        this data and never parses the human message.
      required:
        - code
        - message
      additionalProperties: false
      properties:
        code:
          type: string
          description: >
            Machine-readable, stable, UPPER_SNAKE. Two layers share this field.
            Framework codes, issued by the backend error envelope for every
            non-2xx response including security 401/403: VALIDATION_FAILED
            (400, with fieldErrors), MALFORMED_REQUEST (400), UNAUTHENTICATED
            (401), FORBIDDEN (403), NOT_FOUND (404), METHOD_NOT_ALLOWED (405),
            UNSUPPORTED_MEDIA_TYPE (415), INTERNAL_ERROR (500, generic message
            only). Business codes are declared by each operation when its path
            lands in this file and travel in this same field; a business-rule
            block maps to 409.
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples:
            - VALIDATION_FAILED
            - UNIT_UNAVAILABLE
            - BLOCKED_TRANSITION
        message:
          type: string
          description: Human-readable one-liner (what happened + consequence + next step).
        fieldErrors:
          type: array
          description: Present for validation failures; empty or absent otherwise.
          items:
            $ref: '#/components/schemas/FieldError'
        missingStep:
          type: string
          nullable: true
          description: Machine code for the missing closing step (FR-22 snap-back).
          example: 'RENT_PAYMENT_PENDING'
        stepLabel:
          type: string
          nullable: true
          description: Human-readable closing step name for rendering the snap-back toast.
          example: 'Rent payment is still pending on this check-in.'
    FieldError:
      type: object
      required:
        - field
        - message
      additionalProperties: false
      properties:
        field:
          type: string
          description: camelCase JSON path of the offending field.
        message:
          type: string

    # ------------------------------------------------------------------- list
...
    ContractStatus:
      type: string
      description: Lifecycle status of a contract.
      enum:
        - DRAFT
        - PRINTED
        - SIGNED
        - ACTIVE
        - CLOSED
        - SUPERSEDED
    ContractContentSnapshot:
      type: object
      description: Immutable locked terms snapshot recorded at contract creation.
      required:
        - code
        - reservationCode
        - unitCode
        - monthlyRate
        - baseRent
        - totalRent
        - depositAmount
        - depositRate
        - durationMonths
        - policyVersion
        - currency
      additionalProperties: true
      properties:
        code:
          type: string
          example: 'CT-2026-0001'
        reservationCode:
          type: string
          example: 'BK-2026-0001'
        unitCode:
          type: string
          example: 'M-5'
        monthlyRate:
          type: integer
          example: 690000
        baseRent:
          type: integer
          example: 690000
        totalRent:
          type: integer
          example: 690000
        depositAmount:
          type: integer
          example: 69000
        depositRate:
          type: integer
          example: 10
        durationMonths:
          type: integer
          example: 1
        policyVersion:
          type: string
          example: 'v3'
        currency:
          type: string
          example: 'VND'
        startDate:
          type: string
          format: date
          example: '2026-10-03'
        endDate:
          type: string
          format: date
          example: '2026-11-03'
        customerName:
          type: string
          nullable: true
          example: 'Lan Nguyen'
        customerEmail:
          type: string
          nullable: true
          example: 'lan.nguyen@example.com'
        customerPhone:
          type: string
          nullable: true
          example: '0901234567'
        facilityName:
          type: string
          nullable: true
          example: 'Tan Binh Depot'
        facilityAddress:
          type: string
          nullable: true
          example: '45 Nguyen Van Troi, Tan Binh, Ho Chi Minh City'
        zoneCode:
          type: string
          nullable: true
          example: 'B'
        floor:
          type: integer
          nullable: true
          example: 2
        sizeM2:
          type: number
          format: double
          nullable: true
          example: 8.0
    ContractDto:
      type: object
      description: Full contract representation including locked snapshot terms and revision chaining.
      required:
        - id
        - code
        - reservationId
        - status
        - isLatest
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          example: 1
        code:
          type: string
          example: 'CT-2026-0001'
        reservationId:
          type: integer
          format: int64
          example: 1
        reservationCode:
          type: string
          nullable: true
          example: 'BK-2026-0001'
        policyId:
          type: integer
          nullable: true
          example: 1
        policyVersion:
          type: string
          nullable: true
          example: 'v3'
        contentSnapshot:
          type: string
          description: Raw JSON string of immutable terms snapshot.
        snapshot:
          $ref: '#/components/schemas/ContractContentSnapshot'
        signedPhotoUrl:
          type: string
          nullable: true
          example: '/api/v1/attachments/signed-ct-1.jpg'
        status:
          $ref: '#/components/schemas/ContractStatus'
        supersedesContractId:
          type: integer
          format: int64
          nullable: true
        isLatest:
          type: integer
          example: 1

    # ----------------------------------------------------------------- tasks (3.2)
...
    SignContractRequest:
      type: object
      description: Request payload for attaching signed contract photo.
      required:
        - signedPhotoUrl
      additionalProperties: false
      properties:
        signedPhotoUrl:
          type: string
          example: '/api/v1/attachments/contract-signed-CT-1042-89f4.jpg'
    AttachmentUploadResponseDto:
      type: object
      description: Response containing relative URL and metadata of uploaded attachment.
      required:
        - fileUrl
        - fileName
        - size
        - contentType
      additionalProperties: false
      properties:
        fileUrl:
          type: string
          example: '/api/v1/attachments/contract-signed-CT-1042-89f4.jpg'
        fileName:
          type: string
          example: 'contract-signed-CT-1042-89f4.jpg'
        size:
          type: integer
          format: int64
          example: 204850
        contentType:
          type: string
          example: 'image/jpeg'
...